refactor(parameters): extract arithmetic utilities into arithmetic.go
- Move constrainDifficulty, constrainStability, linearDamping, decayAndFactor,
  initStability, initDifficulty, nextInterval, nextDifficulty, shortTermStability,
  meanReversion, nextRecallStability, nextForgetStability from parameters.go
- Keep parameters.go focused on configuration and high-level NextState API
